SENTRY-2124: LeaderStatusMonitor.toString() throws IllegalFormatConversionException with AtomicLong (Xinran Tinney, reviewed by Sergio Pena)

+++ sentry-provider/sentry-provider-db/src/main/java/org/apache/sentry/provider/db/service/persistent/LeaderStatusMonitor.java
@@ -280,7 +280,7 @@ public final class LeaderStatusMonitor
   public String toString() {
     return isSingleNodeMode?"Leader election disabled":
         String.format("{isSingleNodeMode=%b, incarnationId=%s, isLeader=%b, leaderCount=%d}",
-        isSingleNodeMode, incarnationId, isLeader, leaderCount);
+        isSingleNodeMode, incarnationId, isLeader, leaderCount.longValue());
   }
 
 }
